VolumeUP to take photos
Generally this feature is always extended to both VolUP and VolDOWN
buttons.
But I've applied only to VolumeUP buttons, so VolumeDOWN can be used
traditionally to et show zoom commands, and once the commands are
displayed both volume buttons works as in stock roms.
I think is a good compromise.